+++ packages/twofinger/CHANGELOG.markdown
@@ -4,6 +4,7 @@ Next
 * Fix a dangling section reference in the haddocks.
 * Allow tasty 1.0 along with tasty 0.12.
 * More tests!
+* `Bitraversable1` and `Bifoldable1` for `TwoFingerOddE`.
 
 q4c12-twofinger 0.1 (2018-01-04)
 ================================